Clone each page's content (otherwise we're passing it by reference)
This gem uses the `vulndbhq` gem which has an identity map for each
model type. If we pass the content by reference (as we did before) and
the app downstream modifies the content (e.g. by adding AddonTags to the
content), this trickles back up the chain to the identity map, messing
with subsequent API calls.